feat: move to new components and ensure to memoise
